Modificacion para devolver lista de incidencias en detaller de expediente de modulo de seguimiento

+++ siye-api/src/main/java/mx/com/teclo/siye/persistencia/hibernate/dao/catalogo/PersonaTipoDAOImpl.java
@@ -15,7 +15,7 @@ public class PersonaTipoDAOImpl extends BaseDaoHibernate<PersonaTipoDTO> impleme
 	
 	@SuppressWarnings("unchecked")
 	@Override
-	public List<PersonaTipoDTO> getTecnicos(Integer idTipoPersona) {
+	public List<PersonaTipoDTO> getTecnicos(Long idTipoPersona) {
 		Criteria criteria = getCurrentSession().createCriteria(PersonaTipoDTO.class);
 		criteria.createAlias("tipoPersona", "tipoPersona");
 		criteria.createAlias("persona", "persona");
