+++ siye-api/src/main/java/mx/com/teclo/siye/persistencia/vo/catalogo/PersonaTipoVO.java
@@ -9,13 +9,13 @@ public class PersonaTipoVO implements Serializable {
 	 */
 	private static final long serialVersionUID = 5432701010436407705L;
 	
-	private Integer idPersonaTipo;
+	private Long idPersonaTipo;
 	private PersonaCompVO persona;
 	private TipoPersonaVO tipoPersona;
-	public Integer getIdPersonaTipo() {
+	public Long getIdPersonaTipo() {
 		return idPersonaTipo;
 	}
-	public void setIdPersonaTipo(Integer idPersonaTipo) {
+	public void setIdPersonaTipo(Long idPersonaTipo) {
 		this.idPersonaTipo = idPersonaTipo;
 	}
 	public PersonaCompVO getPersona() {
